server/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
# Теперь используется Together.ai API, а не HuggingFace 